bad decisions
I am continuing the collapse of the import/export layers down to an everything.dart file. It will reduce the number of updates required to make a change ... however ...

It will/may take longer to compile and produce larger binaries if the compiler cannot keep up with the increased number of pointers. However, it is a small application and I have been watching and have seen no signs of either. We will continue for now.

You can import this file anywhere and get access to the entire application. Practice good variable management, good naming of functions and such ... and it will be fine. :-)